requirements.txt
Flask==0.11.1
Flask_Script==2.0.5
pytest==3.0.5
pytest-flask==0.10.0
GitPython==2.1.1
